feat: Add new member dialog to Team component
The code changes include adding a new member dialog to the Team component. This allows users to add new members to the team. The dialog is triggered by clicking on the "Add new member" button and displays a form for entering the member's details. The dialog can be closed by clicking outside the dialog or pressing the escape key.

Note: The recent user commits and recent repository commits are not relevant for generating the commit message in this case.